workspaces/cursor-cli_claude-4.6-opus_20260321T030900Z/cmd/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcube-cli"
version = "0.1.0"
description = "AgentCube command-line interface"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "docker>=6.0.0",
    "rich>=13.0.0",
    "packaging>=23.0",
    "importlib-resources>=6.0.0",
    "semver>=3.0.0",
]

[project.optional-dependencies]
k8s = ["kubernetes>=28.0.0"]

[project.scripts]
kubectl-agentcube = "agentcube.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentcube*"]
